feat: Added support for bidirectional status on status updates
This enables running tasks to be notified of hashes cracked in another task and pause tasks currently running.

We also added support for using a shared directory as a zap synchronization point. If configured, cracked hashes will be written to a directory, and the client will monitor that directory for new zaps.

Signed-off-by: UncleSp1d3r <unclespider@protonmail.com>